Move LinkData and rename it to PaginatedLinkData
Move `LinkData` to the `safir.database` module, since at present
it's only useful for extracting pagination information in the form
used there, and rename it to `PaginatedLinkData`.
