+++ changelog.d/20241122_150037_rra_DM_47769a.md
@@ -1,3 +1,3 @@
 ### New features
 
-- Add new `safir.models.LinkData` model that parses the contents of an HTTP `Link` header and extracts pagination information.
+- Add new `safir.database.PaginatedLinkData` model that parses the contents of an HTTP `Link` header and extracts pagination information.
